requirements.txt
fastapi[standard]
SQLAlchemy
pyjwt
passlib[bcrypt]